Move hooks to hooks folder
