+++ src/components/NostoProduct.tsx
@@ -1,5 +1,4 @@
-import { useNostoContext } from "./context"
-import { useNostoApi } from "../utils/hooks"
+import { useNostoContext, useNostoApi } from "../hooks"
 import { Product } from "../types"
 
 /**
