+++ src/index.ts
@@ -1,2 +1,4 @@
 export type { Cart, Customer, Product, Order, Recommendation } from "./types"
 export * from "./components"
+export { NostoContext, type NostoContextType } from "./context"
+export { useNostoContext } from "./hooks/useNostoContext"
